feat(api): replace validate-urls-from-release with save-urls-from-release
