+++ api/lib/config.js
@@ -87,7 +87,7 @@ export const scheduledJobs = {
   createReleaseTime: process.env.CREATE_RELEASE_TIME,
   exportExternalUrlListTime: process.env.EXPORT_EXTERNAL_URL_LIST_TIME,
   attempts: _getNumber(process.env.CREATE_RELEASE_ATTEMPTS, 4),
-  startCheckUrlJob: isFeatureEnabled(process.env.START_CHECK_URL_JOB),
+  startSaveExternalUrlsJob: isFeatureEnabled(process.env.START_SAVE_EXTERNAL_URLS_JOB ?? process.env.START_CHECK_URL_JOB),
   cleanReleasesTableTime: process.env.CLEAN_RELEASES_TABLE_TIME,
 };
