+++ api/lib/domain/usecases/index.js
@@ -47,4 +47,4 @@ export * from './update-thematic.js';
 export * from './update-tutorial.js';
 export * from './update-tube.js';
 export * from './upload-translation-to-phrase.js';
-export * from './validate-urls-from-release.js';
+export * from './save-urls-from-release.js';
