--- api/lib/infrastructure/scheduled-jobs/check-urls-job-processor.cjs
@@ -1,4 +1,0 @@
-module.exports = async function checkUrlsJobProcessor(job) {
-  const { default: processor } = await import('./check-urls-job-processor.js');
-  return processor(job);
-};
