+++ api/lib/infrastructure/scheduled-jobs/save-external-urls-job-processor.cjs
@@ -1,0 +1,4 @@
+module.exports = async function saveExternalUrlsJobProcessor(job) {
+  const { default: processor } = await import('./save-external-urls-job-processor.js');
+  return processor(job);
+};
