+++ api/lib/infrastructure/scheduled-jobs/save-external-urls-job.js
@@ -4,11 +4,11 @@ import { fileURLToPath } from 'node:url';
 import Queue from 'bull';
 
 const __dirname = fileURLToPath(new URL('.', import.meta.url));
-const cjsFile = __dirname + '/check-urls-job-processor.cjs';
-const esmFile = __dirname + '/check-urls-job-processor.js';
+const cjsFile = __dirname + '/save-external-urls-job-processor.cjs';
+const esmFile = __dirname + '/save-external-urls-job-processor.js';
 
-export async function createCheckUrlsJobQueue() {
-  const queue = createQueue('check-urls-queue');
+export async function createSaveExternalUrlsJobQueue() {
+  const queue = createQueue('save-external-urls-queue');
   if (process.env.NODE_ENV === 'test') {
     const module = await import(esmFile);
     queue.process(module.default);
@@ -26,7 +26,7 @@ const checkUrlsJobOptions = {
 };
 
 export async function start() {
-  const queue = new Queue('check-urls-queue', config.scheduledJobs.redisUrl);
+  const queue = new Queue('save-external-urls-queue', config.scheduledJobs.redisUrl);
   await queue.add({}, checkUrlsJobOptions);
   await queue.close();
 }
