+++ api/tests/tooling/database-builder/factory/index.js
@@ -4,6 +4,7 @@ export * from './build-challenge.js';
 export * from './build-changelog-entry.js';
 export * from './build-competence.js';
 export * from './build-draft-module.js';
+export * from './build-external-url.js';
 export * from './build-framework.js';
 export * from './build-group-from-framework-to-challenge.js';
 export * from './build-localized-challenge.js';
